+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+# v2.0.1
+
+- Add `Delete selection` and `Zoom to` actions in MyWorkspace
+
 # v2.0.0
 
-- Help in reactangle terrain mode only shown in create mode
+- Help in rectangle terrain mode only shown in create mode
 - Add units to ui components
 - Update core and ui to version 6.0